requirements.txt
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
xgboost==2.0.3
ta==0.11.0
praw==7.7.1
requests==2.32.3
python-dotenv==1.0.1
vaderSentiment==3.3.2
alpaca-py==0.26.0
schedule==1.2.2
rich==13.7.1
